Changelog : 1.3
- Added Threading Modes

Single: Download will be executed within the main thread and the GUI
won't respond to your actions if the download isn't finished.

Multiple: Download(s) will be executed with sepereate Threads (QThreads). This is mostly intended for single downloads. You can use that function also for multiple downloads, but that will ruin the progress bar, because it will jump between the different videos.

Your decision ;)