Updated CLI to v2.2
UI:

file function uses the QTreeWidget to let you select from all URLs you want to use.

The same as above for the model / channel function.

Please note, that I can not test that, because I am currently in school and testing with school Wi-Fi is a bit difficult ^^

I test everything in ~ 10 hours and compile v2.3  :)