Now all methods use video objects. Updated to HQPorner V1.1, model function now also works for HQPorner
A lot of the time I reinitialized a Client / Video object again and again because I was to lazy to write a good API for HQPorner. Now where this is done, I can handle video objects a lot better across the UI which gives a good impact for requests usage and makes it more stable.