Added Downloads  to readme.
I  currently kind of restrucutre this project. The repository will be a bit broken (and my commits too) until I am done with it.

V3 will be complete for Android and Desktop, across all OSses.

It's not easy, it will take a lot of time, but it's worth it.

In 2 days school starts again, so not great timing, but could be worse. At least January / February are really easy in Germany.